refactor(v1): rename rollout_log_utils to v1 logging_utils
